requirements.txt
apt-clone==0.2.1
apt-xapian-index==0.47
apturl==0.5.2
astroid==1.4.4
blinker==1.3
boto3==1.4.1
botocore==1.4.78
Brlapi==0.6.4
chardet==2.3.0
colorama==0.3.7
command-not-found==0.3
cryptography==1.2.3
defer==1.0.6
dirspec==13.10
docutils==0.12
flake8==2.5.4
httplib2==0.9.1
idna==2.0
jmespath==0.9.0
lazy-object-proxy==1.2.1
logilab-common==1.1.0
louis==2.6.4
lxml==3.5.0
mccabe==0.2.1
oauthlib==1.0.3
oneconf==0.3.9
pep8==1.7.0
pexpect==4.0.1
Pillow==3.1.2
piston-mini-client==0.7.5
ptyprocess==0.5
pyasn1==0.1.9
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0
pyflakes==1.1.0
pygobject==3.20.0
PyICU==1.9.2
PyJWT==1.3.0
pylint==1.5.2
python-apt==1.1.0b1
python-dateutil==2.6.0
python-debian==0.1.27
python-gnupg==0.3.9
pyxdg==0.25
PyYAML==3.12
reportlab==3.3.0
requests==2.9.1
roman==2.0.0
rsa==3.2.3
s3transfer==0.1.9
screen-resolution-extra==0.0.0
sessioninstaller==0.0.0
six==1.10.0
system-service==0.3
thin-client-config-agent==0.8
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.13.1
virtkey==0.63.0
virtualenv==15.0.1
wrapt==1.8.0
xkit==0.0.0
